# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CredentialStatus": {
        "description": "One credential, as the setup screen sees it. Never carries the value.",
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "effort": {
            "title": "Effort",
            "type": "string"
          },
          "env": {
            "title": "Env",
            "type": "string"
          },
          "group": {
            "title": "Group",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "tail": {
            "title": "Tail",
            "type": "string"
          },
          "unlocks": {
            "title": "Unlocks",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "without_it": {
            "title": "Without It",
            "type": "string"
          }
        },
        "required": [
          "env",
          "label",
          "unlocks",
          "without_it",
          "url",
          "effort",
          "group",
          "required",
          "configured",
          "tail"
        ],
        "title": "CredentialStatus",
        "type": "object"
      },
      "DiagnosticCheck": {
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "fix": {
            "title": "Fix",
            "type": "string"
          },
          "href": {
            "title": "Href",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "level",
          "detail",
          "fix",
          "command",
          "href"
        ],
        "title": "DiagnosticCheck",
        "type": "object"
      },
      "Diagnostics": {
        "properties": {
          "blockers": {
            "title": "Blockers",
            "type": "integer"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticCheck"
            },
            "title": "Checks",
            "type": "array"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "warnings": {
            "title": "Warnings",
            "type": "integer"
          }
        },
        "required": [
          "checks",
          "ready",
          "blockers",
          "warnings"
        ],
        "title": "Diagnostics",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KeyUpdate": {
        "description": "A save. Only the names present here are touched.\n\n`dict[str, str]` rather than a field per credential so that adding one to\n`CREDENTIALS` needs no change here \u2014 the allowlist is `CREDENTIALS` itself,\nchecked at write time, which keeps the two from drifting apart.",
        "properties": {
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Values",
            "type": "object"
          }
        },
        "title": "KeyUpdate",
        "type": "object"
      },
...
      "SetupStatus": {
        "properties": {
          "can_connect": {
            "title": "Can Connect",
            "type": "boolean"
          },
          "can_publish": {
            "title": "Can Publish",
            "type": "boolean"
          },
          "can_render": {
            "title": "Can Render",
            "type": "boolean"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "title": "Channels",
            "type": "array"
          },
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/CredentialStatus"
            },
            "title": "Credentials",
            "type": "array"
          },
          "env_path": {
            "title": "Env Path",
            "type": "string"
          },
          "missing_required": {
            "items": {
              "type": "string"
            },
            "title": "Missing Required",
            "type": "array"
          },
          "worker_running": {
            "title": "Worker Running",
            "type": "boolean"
          }
        },
        "required": [
          "credentials",
          "can_render",
          "can_publish",
          "can_connect",
          "channels",
          "missing_required",
          "env_path",
          "worker_running"
        ],
        "title": "SetupStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Studio Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/setup/diagnostics": {
      "get": {
        "description": "What `scripts/doctor.py` prints, as data.\n\nThe same checks, so the terminal and the screen cannot disagree. It existed\nonly as a script, which meant the answer to \"why did my render fail\" lived\nbehind remembering a virtualenv path \u2014 on the machine of someone who has, by\nconstruction, just failed to set this up.\n\n`network=false` skips the grounding probe, which reaches out to YouTube with a\nsix-second timeout. The Setup screen loads with it off and turns it on for the\nexplicit \"Run checks\" press.",
        "operationId": "diagnostics_v1_setup_diagnostics_get",
        "parameters": [
          {
            "in": "query",
            "name": "network",
            "required": false,
            "schema": {
              "default": true,
              "title": "Network",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagnostics"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diagnostics",
        "tags": [
          "setup"
        ]
      }
    },
    "/v1/setup/keys": {
      "put": {
        "description": "Write credentials to `.env` and make them live in this process.\n\nReturns the new status rather than an acknowledgement, so the screen shows\nwhat is actually in force instead of what it hoped it had set.",
        "operationId": "save_keys_v1_setup_keys_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Keys",
        "tags": [
          "setup"
        ]
      }
    },
...
  }
}